# apis/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/settings": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Возвращает список всех настроек",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Получение списка настроек",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apis.Settings"
                        }
                    },
                    "400": {
                        "description": "Failed to authenticate.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Обновляет указанные настройки",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Обновление настроек",
                "parameters": [
                    {
                        "description": "Данные для обновления настроек",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apis.UpdateSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновление настроек успешно"
                    },
                    "400": {
                        "description": "Failed to authenticate.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/settings/apple/generate-client-secret": {
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Генерирует секретный ключ для использования при авторизации Apple",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Генерация секретного ключа для авторизации Apple",
                "responses": {
                    "200": {
                        "description": "Генерация секретного ключа для авторизации Apple успешно"
                    },
                    "400": {
                        "description": "Failed to authenticate.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/settings/test/email": {
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Проверяет настройки для отправки электронной почты",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Тестирование настроек для электронной почты",
                "parameters": [
                    {
                        "description": "Данные для тестирования настроек электронной почты",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apis.TestEmailSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Тестирование настроек для электронной почты успешно"
                    },
                    "400": {
                        "description": "Failed to authenticate.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/settings/test/s3": {
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Проверяет настройки для хранилища S3",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Тестирование настроек для хранилища S3",
                "parameters": [
                    {
                        "description": "Данные для тестирования настроек S3",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apis.TestS3SettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Тестирование настроек для хранилища S3 успешно"
                    },
                    "400": {
                        "description": "Failed to authenticate.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "apis.AuthProviderConfig": {
            "type": "object",
            "properties": {
                "authUrl": {
                    "type": "string"
                },
                "clientId": {
                    "type": "string"
                },
                "clientSecret": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "tokenUrl": {
                    "type": "string"
                },
                "userApiUrl": {
                    "type": "string"
                }
            }
        },
...
        "apis.BackupsConfig": {
            "type": "object",
            "properties": {
                "cron": {
                    "description": "Cron is a cron expression to schedule auto backups, eg. \"* * * * *\".\n\nLeave it empty to disable the auto backups functionality.",
                    "type": "string"
                },
                "cronMaxKeep": {
                    "description": "CronMaxKeep is the the max number of cron generated backups to\nkeep before removing older entries.\n\nThis field works only when the cron config has valid cron expression.",
                    "type": "integer"
                },
                "s3": {
                    "description": "S3 is an optional S3 storage config specifying where to store the app backups.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/apis.S3Config"
                        }
                    ]
                }
            }
        },
...
        "apis.EmailAuthConfig": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "exceptDomains": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "minPasswordLength": {
                    "type": "integer"
                },
                "onlyDomains": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "apis.EmailTemplate": {
            "type": "object",
            "properties": {
                "actionUrl": {
                    "type": "string"
                },
                "body": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                }
            }
        },
...
        "apis.LogsConfig": {
            "type": "object",
            "properties": {
                "maxDays": {
                    "type": "integer"
                }
            }
        },
...
        "apis.S3Config": {
            "type": "object",
            "properties": {
                "accessKey": {
                    "type": "string"
                },
                "bucket": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "endpoint": {
                    "type": "string"
                },
                "forcePathStyle": {
                    "type": "boolean"
                },
                "region": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                }
            }
        },
...
        "apis.Settings": {
            "type": "object",
            "properties": {
                "adminAuthToken": {
                    "$ref": "#/definitions/apis.TokenConfig"
                },
                "adminFileToken": {
                    "$ref": "#/definitions/apis.TokenConfig"
                },
                "adminPasswordResetToken": {
                    "$ref": "#/definitions/apis.TokenConfig"
                },
                "appleAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "backups": {
                    "$ref": "#/definitions/apis.BackupsConfig"
                },
                "discordAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "emailAuth": {
                    "description": "Deprecated: Will be removed in v0.9+",
                    "allOf": [
                        {
                            "$ref": "#/definitions/apis.EmailAuthConfig"
                        }
                    ]
                },
                "facebookAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "giteaAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "giteeAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "githubAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "gitlabAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "googleAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "kakaoAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "livechatAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "logs": {
                    "$ref": "#/definitions/apis.LogsConfig"
                },
                "meta": {
                    "type": "object",
                    "properties": {
                        "appName": {
                            "type": "string"
                        },
                        "appUrl": {
                            "type": "string"
                        },
                        "confirmEmailChangeTemplate": {
                            "$ref": "#/definitions/apis.EmailTemplate"
                        },
                        "hideControls": {
                            "type": "boolean"
                        },
                        "resetPasswordTemplate": {
                            "$ref": "#/definitions/apis.EmailTemplate"
                        },
                        "senderAddress": {
                            "type": "string"
                        },
                        "senderName": {
                            "type": "string"
                        },
                        "verificationTemplate": {
                            "$ref": "#/definitions/apis.EmailTemplate"
                        }
                    }
                },
                "microsoftAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "oidc2Auth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "oidc3Auth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "oidcAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "recordAuthToken": {
                    "$ref": "#/definitions/apis.TokenConfig"
                },
                "recordEmailChangeToken": {
                    "$ref": "#/definitions/apis.TokenConfig"
                },
                "recordFileToken": {
                    "$ref": "#/definitions/apis.TokenConfig"
                },
                "recordPasswordResetToken": {
                    "$ref": "#/definitions/apis.TokenConfig"
                },
                "recordVerificationToken": {
                    "$ref": "#/definitions/apis.TokenConfig"
                },
                "s3": {
                    "$ref": "#/definitions/apis.S3Config"
                },
                "smtp": {
                    "$ref": "#/definitions/apis.SmtpConfig"
                },
                "spotifyAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "stravaAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "twitchAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "twitterAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                }
            }
        },
        "apis.SmtpConfig": {
            "type": "object",
            "properties": {
                "authMethod": {
                    "description": "SMTP AUTH - PLAIN (default) or LOGIN",
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "host": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "tls": {
                    "description": "Whether to enforce TLS encryption for the mail server connection.\n\nWhen set to false StartTLS command is send, leaving the server\nto decide whether to upgrade the connection or not.",
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "apis.TestEmailSettingsRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "template": {
                    "type": "string"
                }
            }
        },
        "apis.TestS3SettingsRequest": {
            "type": "object",
            "properties": {
                "filesystem": {
                    "description": "The name of the filesystem - storage or backups",
                    "type": "string"
                }
            }
        },
        "apis.TokenConfig": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "secret": {
                    "type": "string"
                }
            }
        },
        "apis.UpdateSettingsRequest": {
            "type": "object",
            "properties": {
                "adminAuthToken": {
                    "$ref": "#/definitions/apis.TokenConfig"
                },
                "adminFileToken": {
                    "$ref": "#/definitions/apis.TokenConfig"
                },
                "adminPasswordResetToken": {
                    "$ref": "#/definitions/apis.TokenConfig"
                },
                "appleAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "backups": {
                    "$ref": "#/definitions/apis.BackupsConfig"
                },
                "discordAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "emailAuth": {
                    "description": "Deprecated: Will be removed in v0.9+",
                    "allOf": [
                        {
                            "$ref": "#/definitions/apis.EmailAuthConfig"
                        }
                    ]
                },
                "facebookAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "giteaAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "giteeAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "githubAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "gitlabAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "googleAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "kakaoAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "livechatAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "logs": {
                    "$ref": "#/definitions/apis.LogsConfig"
                },
                "meta": {
                    "type": "object",
                    "properties": {
                        "appName": {
                            "type": "string"
                        },
                        "appUrl": {
                            "type": "string"
                        },
                        "confirmEmailChangeTemplate": {
                            "$ref": "#/definitions/apis.EmailTemplate"
                        },
                        "hideControls": {
                            "type": "boolean"
                        },
                        "resetPasswordTemplate": {
                            "$ref": "#/definitions/apis.EmailTemplate"
                        },
                        "senderAddress": {
                            "type": "string"
                        },
                        "senderName": {
                            "type": "string"
                        },
                        "verificationTemplate": {
                            "$ref": "#/definitions/apis.EmailTemplate"
                        }
                    }
                },
                "microsoftAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "oidc2Auth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "oidc3Auth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "oidcAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "recordAuthToken": {
                    "$ref": "#/definitions/apis.TokenConfig"
                },
                "recordEmailChangeToken": {
                    "$ref": "#/definitions/apis.TokenConfig"
                },
                "recordFileToken": {
                    "$ref": "#/definitions/apis.TokenConfig"
                },
                "recordPasswordResetToken": {
                    "$ref": "#/definitions/apis.TokenConfig"
                },
                "recordVerificationToken": {
                    "$ref": "#/definitions/apis.TokenConfig"
                },
                "s3": {
                    "$ref": "#/definitions/apis.S3Config"
                },
                "smtp": {
                    "$ref": "#/definitions/apis.SmtpConfig"
                },
                "spotifyAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "stravaAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "twitchAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                },
                "twitterAuth": {
                    "$ref": "#/definitions/apis.AuthProviderConfig"
                }
            }
        }
    }
}